+++ tests/unit/test_run.py
@@ -3480,3 +3480,60 @@ def test_time_as_cftime(scm_run, output_cls):
         assert all([isinstance(v, cftime.DatetimeGregorian) for v in res])
     else:
         assert all([isinstance(v, output_cls) for v in res])
+
+
+def test_round():
+    # There are some quirks with the rounding due to the algo used
+
+    def compare(inp, exp, decimals=1):
+        cols = {
+            "model": "model",
+            "scenario": "scenario",
+            "variable": "variable",
+            "region": "region",
+            "unit": "unit",
+        }
+        index = [2000, 2025, 2030, 2035, 2040]
+        run = ScmRun(data=np.array([inp]).T, index=index, columns=cols,)
+
+        res = run.round(decimals)
+
+        exp = ScmRun(data=np.array([exp]).T, index=index, columns=cols,)
+        assert_scmdf_almost_equal(res, exp)
+
+    compare([3.6565, 5.51, 5.55, 5.45, 1], [3.7, 5.5, 5.6, 5.4, 1.0])
+    compare([-3.6565, -5.51, -5.55, -5.45, -1], [-3.7, -5.5, -5.6, -5.4, -1.0])
+    compare([-3.6565, -5.51, -5.55, -5.45, -1], [-4, -6, -6, -5.0, -1.0], decimals=0)
+    compare([3.6565, 5.51, 5.55, 5.45, 1], [3.6565, 5.51, 5.55, 5.45, 1.0], decimals=5)
+
+
+def test_round_warns_small():
+    run = ScmRun(
+        data=np.array([[3.6565, 5.51, 5.55, 1, 2.34e-2]]).T,
+        index=[2000, 2025, 2030, 2035, 2040],
+        columns={
+            "model": "model",
+            "scenario": "scenario",
+            "variable": "variable",
+            "region": "region",
+            "unit": "unit",
+        },
+    )
+
+    match = "There are small values which may be truncated during rounding"
+
+    with pytest.warns(UserWarning, match=match):
+        res = run.round(1)
+
+    exp = ScmRun(
+        data=np.array([[3.7, 5.5, 5.6, 1.0, 0]]).T,
+        index=[2000, 2025, 2030, 2035, 2040],
+        columns={
+            "model": "model",
+            "scenario": "scenario",
+            "variable": "variable",
+            "region": "region",
+            "unit": "unit",
+        },
+    )
+    assert_scmdf_almost_equal(res, exp)